+++ src/pages/meeting/MeetingPlaceAdd.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { PlaceType } from "components/common/card/DisplayListDetailCard"
+import { PlaceType } from "components/common/card/PlaceDetailCard "
 import SearchPlace from "pages/course/SearchPlace"
 
 const MeetingPlaceAdd = (): JSX.Element => {
