+++ src/types/API/course-service.ts
@@ -1,5 +1,17 @@
 // Course-Service API Docs 참고
 
+export interface SearchCardProp {
+  index: number // 카드의 인덱스 넘버 - order
+  address_name: string // 주소
+  category_name: string // 플레이스 카테고리 -placeCategory
+  place_name: string // 장소 이름           -name
+  place_url: string // 플레이스 주소        -
+  x: number // 경도 longitude              -lon
+  y: number // 위도 latitude               -lat
+  description: string // 설명              -description
+  id: number // 카카오 id          -kakaoPlaceId
+}
+
 export interface CourseData {
   title: string
   description: string
